requirements.txt
fastapi
pydantic[dotenv]
uvicorn
python-dotenv
langgraph
langsmith
langchain_community
duckduckgo-search
langchain-openai
langchain-tavily
langchain-groq
langchain-google-community[calendar]
pytz